+++ 21H2/Wdf01000/-Assign@FxString@@QEAAJPEBG@Z.c
@@ -1,13 +1,13 @@
 /*
- * XREFs of ?Assign@FxString@@QEAAJPEBG@Z @ 0x1C002D548
+ * XREFs of ?Assign@FxString@@QEAAJPEBG@Z @ 0x1C0061ACC
  * Callers:
- *     imp_WdfRegistryQueryMultiString @ 0x1C002D190 (imp_WdfRegistryQueryMultiString.c)
- *     imp_WdfDriverRetrieveVersionString @ 0x1C0066830 (imp_WdfDriverRetrieveVersionString.c)
+ *     imp_WdfDriverRetrieveVersionString @ 0x1C004C5F0 (imp_WdfDriverRetrieveVersionString.c)
+ *     imp_WdfRegistryQueryMultiString @ 0x1C005E500 (imp_WdfRegistryQueryMultiString.c)
  * Callees:
- *     ?FxDuplicateUnicodeString@@YAJPEAU_FX_DRIVER_GLOBALS@@PEBU_UNICODE_STRING@@PEAU2@@Z @ 0x1C00273A8 (-FxDuplicateUnicodeString@@YAJPEAU_FX_DRIVER_GLOBALS@@PEBU_UNICODE_STRING@@PEAU2@@Z.c)
+ *     ?FxDuplicateUnicodeString@@YAJPEAU_FX_DRIVER_GLOBALS@@PEBU_UNICODE_STRING@@PEAU2@@Z @ 0x1C0061C1C (-FxDuplicateUnicodeString@@YAJPEAU_FX_DRIVER_GLOBALS@@PEBU_UNICODE_STRING@@PEAU2@@Z.c)
  */
 
-__int64 __fastcall FxString::Assign(FxString *this, const wchar_t *SourceString)
+int __fastcall FxString::Assign(FxString *this, const wchar_t *SourceString)
 {
   _UNICODE_STRING string; // [rsp+20h] [rbp-18h] BYREF
