+++ 21H2/Wdf01000/-_DispatchWaitWake@FxPkgPnp@@KAJPEAV1@PEAVFxIrp@@@Z.c
@@ -1,12 +1,12 @@
 /*
- * XREFs of ?_DispatchWaitWake@FxPkgPnp@@KAJPEAV1@PEAVFxIrp@@@Z @ 0x1C0039850
+ * XREFs of ?_DispatchWaitWake@FxPkgPnp@@KAJPEAV1@PEAVFxIrp@@@Z @ 0x1C0083050
  * Callers:
  *     <none>
  * Callees:
- *     ?DispatchWaitWake@FxPkgPnp@@IEAAJPEAVFxIrp@@@Z @ 0x1C0088098 (-DispatchWaitWake@FxPkgPnp@@IEAAJPEAVFxIrp@@@Z.c)
+ *     ?DispatchWaitWake@FxPkgPnp@@IEAAJPEAVFxIrp@@@Z @ 0x1C0080034 (-DispatchWaitWake@FxPkgPnp@@IEAAJPEAVFxIrp@@@Z.c)
  */
 
-int __fastcall FxPkgPnp::_DispatchWaitWake(FxPkgPnp *This, FxIrp *Irp)
+__int64 __fastcall FxPkgPnp::_DispatchWaitWake(FxPkgPnp *This, FxIrp *Irp)
 {
   return FxPkgPnp::DispatchWaitWake(This, Irp);
 }
